Move to float from long for mass sample data point storage.

+++ endpoint/arduino/src/lib/eeprom.h
@@ -5,7 +5,7 @@
 
 struct EEPROMStruct {
     float scale;
-    int32_t offset[MAX_NUMBER_OF_MASS_SENSORS];
+    float offset[MAX_NUMBER_OF_MASS_SENSORS];
 };
 
 #endif /* eeprom_h */
