+++ endpoint/arduino/src/utils/sort.cpp
@@ -1,8 +1,8 @@
 #include "sort.h"
 
-void insertion_sort(long* arr, int size) {
+void insertion_sort(float* arr, int size) {
     for (int i = 1; i < size; ++i) {
-        long key = arr[i];
+        float key = arr[i];
         int j = i - 1;
         while (j >= 0 && arr[j] > key) {
             arr[j + 1] = arr[j];
